Fix bug delete extra item
This approach seems to work for correctly opening and closing the project content with a smooth transition.
To be able to drop an item with the project closed, an implementation similar to the Quick component could be a valid approach. This would go in an {:else} with another Sortable in the IncubatorProject component.